handgesture-rasp/requirements.txt
imutils==0.5.2
matplotlib==3.0.3
numpy==1.16.2
opencv-contrib-python==3.4.1.15
opencv-python==3.4.1.15
Pillow==6.0.0
pyserial==3.4
scipy==1.2.1
tensorflow==1.13.1
